+++ apps/server/src/shared/filters/app-exception.filter.ts
@@ -81,7 +81,6 @@ export class AppExceptionFilter implements ExceptionFilter {
       };
     } else if (exception instanceof Error) {
       this.logger.error(exception.stack ?? exception.message);
-      body.error.message = exception.message;
     } else {
       this.logger.error(String(exception));
     }